+++ extensions/tags-core/src/main/kotlin/io/medatarun/tags/core/infra/db/tables/TagGroupTable.kt
@@ -2,7 +2,7 @@ package io.medatarun.tags.core.infra.db.tables
 
 import org.jetbrains.exposed.v1.core.Table
 
-internal object TagGroupTable : Table("tag_group") {
+internal object TagGroupTable : Table("tag_group_projection") {
     val id = text("id")
     val key = text("key")
     val name = text("name").nullable()